refactor: move all playgrounds as playground subfolders
