+++ .eslintignore
@@ -1,10 +1,10 @@
 demos
 build
-playground
+playground/index.html
+playground/angular
 node_modules
 dist
 package
 src/swiper.js
 src/angular
-playground-angular
 postinstall.js
