+++ angular.json
@@ -80,29 +80,29 @@
           "skipTests": true
         }
       },
-      "root": "playground-angular",
-      "sourceRoot": "playground-angular",
+      "root": "playground/angular",
+      "sourceRoot": "playground/angular",
       "prefix": "app",
       "architect": {
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
           "options": {
             "outputPath": "dist/demo/browser",
-            "index": "playground-angular/src/index.html",
-            "main": "playground-angular/src/main.ts",
-            "polyfills": "playground-angular/src/polyfills.ts",
-            "tsConfig": "playground-angular/tsconfig.app.json",
+            "index": "playground/angular/src/index.html",
+            "main": "playground/angular/src/main.ts",
+            "polyfills": "playground/angular/src/polyfills.ts",
+            "tsConfig": "playground/angular/tsconfig.app.json",
             "aot": true,
-            "assets": ["playground-angular/src/favicon.ico", "playground-angular/src/assets"],
-            "styles": ["playground-angular/src/styles.scss"],
+            "assets": ["playground/angular/src/favicon.ico", "playground/angular/src/assets"],
+            "styles": ["playground/angular/src/styles.scss"],
             "scripts": []
           },
           "configurations": {
             "production": {
               "fileReplacements": [
                 {
-                  "replace": "playground-angular/src/environments/environment.ts",
-                  "with": "playground-angular/src/environments/environment.prod.ts"
+                  "replace": "playground/angular/src/environments/environment.ts",
+                  "with": "playground/angular/src/environments/environment.prod.ts"
                 }
               ],
               "optimization": true,
@@ -147,12 +147,12 @@
         "test": {
           "builder": "@angular-devkit/build-angular:karma",
           "options": {
-            "main": "playground-angular/src/test.ts",
-            "polyfills": "playground-angular/src/polyfills.ts",
-            "tsConfig": "playground-angular/tsconfig.spec.json",
-            "karmaConfig": "playground-angular/karma.conf.js",
-            "assets": ["playground-angular/src/favicon.ico", "playground-angular/src/assets"],
-            "styles": ["playground-angular/src/styles.scss"],
+            "main": "playground/angular/src/test.ts",
+            "polyfills": "playground/angular/src/polyfills.ts",
+            "tsConfig": "playground/angular/tsconfig.spec.json",
+            "karmaConfig": "playground/angular/karma.conf.js",
+            "assets": ["playground/angular/src/favicon.ico", "playground/angular/src/assets"],
+            "styles": ["playground/angular/src/styles.scss"],
             "scripts": []
           }
         },
@@ -160,8 +160,8 @@
           "builder": "@angular-devkit/build-angular:tslint",
           "options": {
             "tsConfig": [
-              "playground-angular/tsconfig.app.json",
-              "playground-angular/tsconfig.server.json"
+              "playground/angular/tsconfig.app.json",
+              "playground/angular/tsconfig.server.json"
             ],
             "exclude": ["**/node_modules/**"]
           }
@@ -170,16 +170,16 @@
           "builder": "@angular-devkit/build-angular:server",
           "options": {
             "outputPath": "dist/demo/server",
-            "main": "playground-angular/server.ts",
-            "tsConfig": "playground-angular/tsconfig.server.json"
+            "main": "playground/angular/server.ts",
+            "tsConfig": "playground/angular/tsconfig.server.json"
           },
           "configurations": {
             "production": {
               "outputHashing": "media",
               "fileReplacements": [
                 {
-                  "replace": "playground-angular/src/environments/environment.ts",
-                  "with": "playground-angular/src/environments/environment.prod.ts"
+                  "replace": "playground/angular/src/environments/environment.ts",
+                  "with": "playground/angular/src/environments/environment.prod.ts"
                 }
               ],
               "sourceMap": false,
