+++ playground/angular/e2e/tsconfig.json
