+++ playground/angular/src/app/app.component.scss
