+++ playground/react/main.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 // eslint-disable-next-line
-import '../build/swiper-bundle.css';
+import '../../build/swiper-bundle.css';
 
 import App from './App.jsx';
