+++ playground/vue/main.js
@@ -1,7 +1,7 @@
 import { createApp } from 'vue';
 
 // eslint-disable-next-line
-import '../build/swiper-bundle.css';
+import '../../build/swiper-bundle.css';
 
 import App from './App.vue';
